refactor: organize hooks with barrel exports
- Move useCustomToast and useOnboarding to src/hooks/common/
- Create barrel exports for common, auth, and api hooks
- Add main hooks index file for centralized exports
- Update all imports throughout the application
- Fix useCustomToast to export both default and named exports
- Update useOnboarding to use correct auth import
