+++ frontend/src/components/QuizCreation/DeleteQuizConfirmation.tsx
@@ -16,7 +16,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog"
-import useCustomToast from "@/hooks/useCustomToast"
+import { useCustomToast } from "@/hooks/common"
 import { handleError } from "@/utils"
 
 interface DeleteQuizConfirmationProps {
