+++ frontend/src/components/UserSettings/DeleteConfirmation.tsx
@@ -16,7 +16,7 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog"
 import useAuth from "@/hooks/useCanvasAuth"
-import useCustomToast from "@/hooks/useCustomToast"
+import { useCustomToast } from "@/hooks/common"
 import { handleError } from "@/utils"
 
 const DeleteConfirmation = () => {
